pyproject.toml
[project]
name = "nijisanji-stream-playlist-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "janome",
    "lingua-language-detector",
    "oauth2client",
    "pyyaml",
    "requests",
    "selenium",
]
